Drastically simplify source caching
Removed configurable procedure identifiers, because MD5'ing the method
signature was the only reasonable setting anyways. Now we don't need to
rename files when the setting changes.

Separated the two distinct "states" of Procedure into ReifiedProcedure
and SkeletonProcedure. The skeleton stores only the metadata about the
procedure, which is all insignificant when detecting if a procedure's
body changed (requiring a recompile). ReifiedProcedure contains the
actual source code and is compared to the skeleton's disk-cached source
code. This simplifies all sorts of weird statefulness issues